polish code for qgpo
