+++ ding/example/qgpo_halfcheetah_medium_expert.py
@@ -6,9 +6,6 @@ import treetensor.torch as ttorch
 import gym
 import d4rl
 import numpy as np
-import sklearn
-import sklearn.datasets
-from sklearn.utils import shuffle as util_shuffle
 import tqdm
 from easydict import EasyDict
 
@@ -25,163 +22,6 @@ from ding.framework.middleware import interaction_evaluator, trainer, CkptSaver,
 from ding.framework.middleware.functional.evaluator import VectorEvalMonitor
 from ding.utils import set_pkg_seed
 from ding.torch_utils import to_ndarray, get_shape0
-
-
-# Dataset iterator
-def inf_train_gen(data, batch_size=200):
-    print(data)
-    if data == "swissroll":
-        print(data)
-        data = sklearn.datasets.make_swiss_roll(n_samples=batch_size, noise=1.0)[0]
-        data = data.astype("float32")[:, [0, 2]]
-        data /= 5
-        return data, np.sum(data ** 2, axis=-1, keepdims=True) / 9.0
-    elif data == "circles":
-        data = sklearn.datasets.make_circles(n_samples=batch_size, factor=.5, noise=0.08)[0]
-        data = data.astype("float32")
-        data *= 3
-        return data
-    elif data == "rings":
-        n_samples4 = n_samples3 = n_samples2 = batch_size // 4
-        n_samples1 = batch_size - n_samples4 - n_samples3 - n_samples2
-
-        # so as not to have the first point = last point, we set endpoint=False
-        linspace4 = np.linspace(0, 2 * np.pi, n_samples4, endpoint=False)
-        linspace3 = np.linspace(0, 2 * np.pi, n_samples3, endpoint=False)
-        linspace2 = np.linspace(0, 2 * np.pi, n_samples2, endpoint=False)
-        linspace1 = np.linspace(0, 2 * np.pi, n_samples1, endpoint=False)
-
-        circ4_x = np.cos(linspace4)
-        circ4_y = np.sin(linspace4)
-        circ3_x = np.cos(linspace4) * 0.75
-        circ3_y = np.sin(linspace3) * 0.75
-        circ2_x = np.cos(linspace2) * 0.5
-        circ2_y = np.sin(linspace2) * 0.5
-        circ1_x = np.cos(linspace1) * 0.25
-        circ1_y = np.sin(linspace1) * 0.25
-
-        X = np.vstack(
-            [np.hstack([circ4_x, circ3_x, circ2_x, circ1_x]),
-             np.hstack([circ4_y, circ3_y, circ2_y, circ1_y])]
-        ).T * 3.0
-        X = util_shuffle(X)
-
-        center_dist = X[:, 0] ** 2 + X[:, 1] ** 2
-        energy = np.zeros_like(center_dist)
-
-        energy[(center_dist >= 8.5)] = 0.667
-        energy[(center_dist >= 5.0) & (center_dist < 8.5)] = 0.333
-        energy[(center_dist >= 2.0) & (center_dist < 5.0)] = 1.0
-        energy[(center_dist < 2.0)] = 0.0
-
-        # Add noise
-        X = X + np.random.normal(scale=0.08, size=X.shape)
-
-        return X.astype("float32"), energy[:, None]
-
-    elif data == "moons":
-        data, y = sklearn.datasets.make_moons(n_samples=batch_size, noise=0.1)
-        data = data.astype("float32")
-        data = data * 2 + np.array([-1, -0.2])
-        return data.astype(np.float32), (y > 0.5).astype(np.float32)[:, None]
-
-    elif data == "8gaussians":
-        scale = 4.
-        centers = [
-            (0, 1),
-            (-1. / np.sqrt(2), 1. / np.sqrt(2)),
-            (-1, 0),
-            (-1. / np.sqrt(2), -1. / np.sqrt(2)),
-            (0, -1),
-            (1. / np.sqrt(2), -1. / np.sqrt(2)),
-            (1, 0),
-            (1. / np.sqrt(2), 1. / np.sqrt(2)),
-        ]
-
-        centers = [(scale * x, scale * y) for x, y in centers]
-
-        dataset = []
-        indexes = []
-        for i in range(batch_size):
-            point = np.random.randn(2) * 0.5
-            idx = np.random.randint(8)
-            center = centers[idx]
-            point[0] += center[0]
-            point[1] += center[1]
-            indexes.append(idx)
-            dataset.append(point)
-        dataset = np.array(dataset, dtype="float32")
-        dataset /= 1.414
-        return dataset, np.array(indexes, dtype="float32")[:, None] / 7.0
-
-    elif data == "pinwheel":
-        radial_std = 0.3
-        tangential_std = 0.1
-        num_classes = 5
-        num_per_class = batch_size // 5
-        rate = 0.25
-        rads = np.linspace(0, 2 * np.pi, num_classes, endpoint=False)
-
-        features = np.random.randn(num_classes*num_per_class, 2) \
-            * np.array([radial_std, tangential_std])
-        features[:, 0] += 1.
-        labels = np.repeat(np.arange(num_classes), num_per_class)
-
-        angles = rads[labels] + rate * np.exp(features[:, 0])
-        rotations = np.stack([np.cos(angles), -np.sin(angles), np.sin(angles), np.cos(angles)])
-        rotations = np.reshape(rotations.T, (-1, 2, 2))
-
-        return 2 * np.random.permutation(np.einsum("ti,tij->tj", features, rotations))
-
-    elif data == "2spirals":
-        n = np.sqrt(np.random.rand(batch_size // 2, 1)) * 540 * (2 * np.pi) / 360
-        d1x = -np.cos(n) * n + np.random.rand(batch_size // 2, 1) * 0.5
-        d1y = np.sin(n) * n + np.random.rand(batch_size // 2, 1) * 0.5
-        x = np.vstack((np.hstack((d1x, d1y)), np.hstack((-d1x, -d1y)))) / 3
-        x += np.random.randn(*x.shape) * 0.1
-        return x, np.clip((1 - np.concatenate([n, n]) / 10), 0, 1)
-
-    elif data == "checkerboard":
-        x1 = np.random.rand(batch_size) * 4 - 2
-        x2_ = np.random.rand(batch_size) - np.random.randint(0, 2, batch_size) * 2
-        x2 = x2_ + (np.floor(x1) % 2)
-        points = np.concatenate([x1[:, None], x2[:, None]], 1) * 2
-
-        points_x = points[:, 0]
-        judger = ((points_x > 0) & (points_x <= 2)) | ((points_x <= -2))
-        return points, judger.astype(np.float32)[:, None]
-
-    elif data == "line":
-        x = np.random.rand(batch_size) * 5 - 2.5
-        y = x
-        return np.stack((x, y), 1)
-    elif data == "cos":
-        x = np.random.rand(batch_size) * 5 - 2.5
-        y = np.sin(x) * 2.5
-        return np.stack((x, y), 1)
-    else:
-        assert False
-
-
-class Toy_dataset(torch.utils.data.Dataset):
-
-    def __init__(self, name, datanum=1000000):
-        assert name in ["swissroll", "8gaussians", "moons", "rings", "checkerboard", "2spirals"]
-        self.datanum = datanum
-        self.name = name
-        self.datas, self.energy = inf_train_gen(name, batch_size=datanum)
-        self.datas = torch.Tensor(self.datas).to("cuda")
-        self.energy = torch.Tensor(self.energy).to("cuda")
-        self.datadim = 2
-
-    def __getitem__(self, index):
-        return {"a": self.datas[index], "e": self.energy[index]}
-
-    def __add__(self, other):
-        raise NotImplementedError
-
-    def __len__(self):
-        return self.datanum
 
 
 def return_range(dataset, max_episode_steps):
@@ -268,7 +108,6 @@ main_config = dict(
         cuda=True,
         on_policy=False,
         #load_path='./halfcheetah_medium_expert_v2_QGPO_seed0/ckpt/iteration_600000.pth.tar',
-        #load_path='./halfcheetah_medium_expert_v2_QGPO_seed0_231018_001550/ckpt/iteration_500000.pth.tar',
         model=dict(
             score_net=dict(
                 device='cuda',
@@ -305,27 +144,6 @@ main_config = dict(
     ),
 )
 create_config = dict()
-
-# TASK="halfcheetah-medium-expert-v2";
-# seed=0;
-# setting="reproduce";
-# python3
-# -u train_behavior.py
-# --expid $TASK${seed}${setting}
-# --env $TASK
-# --seed ${seed}
-
-# TASK="halfcheetah-medium-expert-v2"; seed=0; setting="reproduce";
-# python3
-# -u train_critic.py
-# --actor_load_path models_rl/halfcheetah-medium-expert-v20reproduce/behavior_ckpt600.pth
-# --expid $TASK${seed}${setting}
-# --env $TASK
-# --diffusion_steps 15
-# --seed ${seed}
-# --alpha 3
-# --q_alpha 1
-# --method "CEP"
 
 
 def QGPO_support_data_generator(cfg, dataset, policy) -> Callable:
@@ -390,8 +208,6 @@ def QGPO_support_data_generator(cfg, dataset, policy) -> Callable:
     return _data_generator
 
 
-
-
 def interaction_qgpo_evaluator(cfg: EasyDict, policy: Policy, env: BaseEnvManager, render: bool = False) -> Callable:
     """
     Overview:
@@ -424,7 +240,7 @@ def interaction_qgpo_evaluator(cfg: EasyDict, policy: Policy, env: BaseEnvManage
            (ctx.train_iter - ctx.last_eval_iter < cfg.policy.eval.evaluator.eval_freq):
             return
 
-        ctx.info_for_logging={}
+        ctx.info_for_logging = {}
 
         for guidance_scale in cfg.policy.eval.guidance_scale:
 
@@ -469,14 +285,21 @@ def interaction_qgpo_evaluator(cfg: EasyDict, policy: Policy, env: BaseEnvManage
                     )
                 )
             elif isinstance(ctx, OfflineRLContext):
-                logging.info('Evaluation: Train Iter({})\tEval Reward({:.3f})\tguidance_scale({})'.format(ctx.train_iter, episode_return, guidance_scale))
+                logging.info(
+                    'Evaluation: Train Iter({})\tEval Reward({:.3f})\tguidance_scale({})'.format(
+                        ctx.train_iter, episode_return, guidance_scale
+                    )
+                )
             else:
                 raise TypeError("not supported ctx type: {}".format(type(ctx)))
             ctx.last_eval_iter = ctx.train_iter
             ctx.eval_value = episode_return
-            ctx.eval_value_min = min(episode_return_min, ctx.eval_value_min) if hasattr(ctx, 'eval_value_min') else episode_return_min
-            ctx.eval_value_max = max(episode_return_max, ctx.eval_value_max) if hasattr(ctx, 'eval_value_max') else episode_return_max
-            ctx.eval_value_std = max(episode_return_std, ctx.eval_value_std) if hasattr(ctx, 'eval_value_std') else episode_return_std
+            ctx.eval_value_min = min(episode_return_min,
+                                     ctx.eval_value_min) if hasattr(ctx, 'eval_value_min') else episode_return_min
+            ctx.eval_value_max = max(episode_return_max,
+                                     ctx.eval_value_max) if hasattr(ctx, 'eval_value_max') else episode_return_max
+            ctx.eval_value_std = max(episode_return_std,
+                                     ctx.eval_value_std) if hasattr(ctx, 'eval_value_std') else episode_return_std
             ctx.last_eval_value = ctx.eval_value
             ctx.eval_output = {'episode_return': episode_return}
             episode_info = eval_monitor.get_episode_info()
@@ -530,13 +353,13 @@ def qgpo_offline_data_fetcher(cfg: EasyDict, dataset: Dataset, collate_fn=lambda
         while True:
             yield from dataloader
 
-    data=get_behavior_policy_training_data()
+    data = get_behavior_policy_training_data()
 
     def get_q_training_data():
         while True:
             yield from dataloader_q
 
-    data_q=get_q_training_data()
+    data_q = get_q_training_data()
 
     def _fetch(ctx: "OfflineRLContext"):
         """
@@ -560,8 +383,6 @@ def qgpo_offline_data_fetcher(cfg: EasyDict, dataset: Dataset, collate_fn=lambda
     return _fetch
 
 
-
-
 def main():
     # If you don't have offline data, you need to prepare if first and set the data_path in config
     # For demostration, we also can train a RL policy (e.g. SAC) and collect some data
@@ -574,7 +395,7 @@ def main():
 
         model = QGPO(cfg=cfg.policy.model)
         policy = QGPOPolicy(cfg.policy, model=model)
-        if hasattr(cfg.policy,"load_path") and cfg.policy.load_path is not None:
+        if hasattr(cfg.policy, "load_path") and cfg.policy.load_path is not None:
             policy_state_dict = torch.load(cfg.policy.load_path, map_location=torch.device("cpu"))
             policy.learn_mode.load_state_dict(policy_state_dict)
 
@@ -585,28 +406,30 @@ def main():
             ],
             cfg=cfg.env.manager
         )
-        #task.use(epoch_timer())
+
         task.use(QGPO_support_data_generator(cfg, dataset, policy))
         task.use(qgpo_offline_data_fetcher(cfg, dataset, collate_fn=None))
         task.use(trainer(cfg, policy.learn_mode))
         task.use(interaction_qgpo_evaluator(cfg, policy.eval_mode, evaluator_env))
-        task.use(wandb_offline_logger(
-            cfg = EasyDict(
-                dict(
-                    gradient_logger=False,
-                    plot_logger=True,
-                    video_logger=False,
-                    action_logger=False,
-                    return_logger=False,
-                    vis_dataset=False,
-                )
-            ),
-            exp_config=cfg,
-            project_name=cfg.exp_name)
+        task.use(
+            wandb_offline_logger(
+                cfg=EasyDict(
+                    dict(
+                        gradient_logger=False,
+                        plot_logger=True,
+                        video_logger=False,
+                        action_logger=False,
+                        return_logger=False,
+                        vis_dataset=False,
+                    )
+                ),
+                exp_config=cfg,
+                project_name=cfg.exp_name
+            )
         )
         task.use(CkptSaver(policy, cfg.exp_name, train_freq=100000))
         task.use(offline_logger())
-        task.use(termination_checker(max_train_iter=500000+cfg.policy.learn.q_value_stop_training_iter))
+        task.use(termination_checker(max_train_iter=500000 + cfg.policy.learn.q_value_stop_training_iter))
         task.run()
